+++ src/dtos/chat.dto.ts
@@ -1,0 +1,7 @@
+export class CreateChatDto {
+    readonly name: string;
+    readonly type: string;
+    readonly description: string;
+    readonly messageIds: number[];
+    readonly userIds: number[];
+}