+++ src/dtos/user.dto.ts
@@ -1,0 +1,7 @@
+export class CreateUserDto {
+    readonly firstName: string;
+    readonly lastName: string;
+    readonly nickName: string;
+    readonly phoneNumber: string;
+    readonly messageIds: number[];
+}