+++ src/modules/index.ts
@@ -1,0 +1,3 @@
+export * from './chat.module';
+export * from './message.module';
+export * from './user.module';